requirements.txt
langchain>=0.1.17
langchain-community>=0.0.24
openai>=1.3.5
faiss-cpu
flask==2.2.5
python-dotenv==1.0.1
twilio==8.10.1
pymongo==4.6.1


